+++ libs/application/templates/signature-collection/municipal-list-creation/src/forms/Draft.ts
@@ -1,6 +1,4 @@
-import {
-  buildForm,
-} from '@island.is/application/core'
+import { buildForm } from '@island.is/application/core'
 import { Form, FormModes } from '@island.is/application/types'
 import Logo from '../../assets/Logo'
 
@@ -14,9 +12,5 @@ export const Draft: Form = buildForm({
   renderLastScreenButton: true,
   renderLastScreenBackButton: true,
   logo: Logo,
-  children: [
-    information,
-    overview,
-    confirmation
-  ],
+  children: [information, overview, confirmation],
 })
